backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
pymongo==4.9.2
dnspython==2.6.1
google-generativeai==0.8.3
openai==1.51.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
email-validator==2.2.0
httpx==0.27.2
celery==5.4.0
redis==5.1.1
